requirements.txt
fastapi[standard]
python-multipart
uvicorn[standard]
pydantic
langchain_community
python-jose[cryptography]
passlib[bcrypt]
sqlmodel